requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
pymysql==0.9.3
pytest==5.4.2
pandas== 1.1.5
botocore==1.15.32
boto3==1.12.32
matplotlib==3.3.4
numpy== 1.19.5
sklearn==0.0



